Update types and readme
