+++ types/src/cli.d.ts
@@ -1,0 +1,9 @@
+#!/usr/bin/env node
+export type Arguments = {
+    _: Array<string>;
+    input: string;
+    output: string;
+    props: string;
+};
+declare const props: any;
+export {};
