+++ types/src/index.d.ts
@@ -1,9 +1,3 @@
 #!/usr/bin/env node
-export type Arguments = {
-    _: Array<string>;
-    input: string;
-    output: string;
-    props: string;
-};
-declare const props: any;
-export {};
+export { compile };
+import { compile } from "./compiler";
